refact: 컴포넌트 별로 코드 분리, 주석 추가
